+++ visualization/include/pcl/visualization/common/ren_win_interact_map.h
@@ -38,6 +38,8 @@
 
 #pragma once
 
+#include <pcl/pcl_exports.h>
+
 #include <map>
 #include <string>
 
@@ -53,7 +55,7 @@ namespace pcl
 {
   namespace visualization
   {
-    class RenWinInteract
+    class PCL_EXPORTS RenWinInteract
     {
       public:
